Begin updating app usage for new alternate app icon resources
- Improve default icon for clear mode
